backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
python-decouple==3.8
psycopg2-binary==2.9.10
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
requests==2.31.0
celery==5.3.4
redis==5.0.3
gunicorn==21.2.0
django-filter==24.2
Pillow==10.3.0
